# Optimize WAL segment reading