###### Required for all non-trivial PRs
- [x] Rebased/mergable
- [x] Tests pass
- [x] CHANGELOG.md updated
- [ ] Sign [CLA](https://influxdata.com/community/cla/) (if not already signed)

Reading WAL segments generates a lot of garbage and was doing some things inefficiently.  This change improves reading segments a bit.  

These changes help a little, but most of the garbage (and time) is still dominated by `tsm1.Value` interface and conversion when re-creating the type specific values to be added back to the `tsm1.Cache`.  Fixing that is a bit more complicated.

```
benchmark                       old ns/op     new ns/op     delta
BenchmarkWALSegmentReader-8     40012977      23665143      -40.86%

benchmark                       old allocs     new allocs     delta
BenchmarkWALSegmentReader-8     1000500        500601         -49.96%

benchmark                       old bytes     new bytes     delta
BenchmarkWALSegmentReader-8     24236300      16236804      -33.01%
```